chore(repo): restructure for clarity and remove root clutter
Group documentation into docs/{architecture,audits,planning,retrospectives,dev-tools}/,
rename the spaced 'STATIC DATA/' folder to data/static/, relocate legacy assets under
data/legacy/ and scripts/legacy/, move the misplaced src/data/sampleData.ts into
src/lib/, and fix the 'depracated' typo folder. Update all runtime, script, pipeline,
and test references to match the new paths; rewrite README.md around the new layout.

- Root cleanup: audit/planning .md files moved out of the project root
- data/: static/ for curated source lists, legacy/ for historical/deprecated data
- scripts/: legacy/ for retired generators
- Fixes path references in src/lib/data.ts, config/generationPipeline.config.ts,
  scripts/{analyzeWordsAndSentences,buildSyntheticWordPracticeData,generateWordAudio}.ts
- Ignore .vite/ build cache
- Tests: 126/126 pass; vite build succeeds
