+++ .gitignore
@@ -22,6 +22,7 @@ Thumbs.db
 node_modules/
 dist/
 *.tsbuildinfo
+.vite/
 
 # Speech debug / temp artifacts
 data/debug/
