+++ docs/architecture/performance-notes.md
