+++ docs/architecture/routes.md
