+++ docs/audits/SECURITY_AUDIT_AND_HARDENING.md
