+++ scripts/legacy/generateSentences.legacy.py
