+++ src/types/contentGeneration.ts
@@ -109,7 +109,7 @@ export interface GenerationConfigVoice {
  * Defines all input/output paths for words, sentences, audio, and indices.
  */
 export interface GenerationPathsConfig {
-  rawWordsJsonPath: string;      // e.g., "STATIC DATA/words.json"
+  rawWordsJsonPath: string;      // e.g., "data/static/words.json"
   rawSentencesJsonPath: string | string[];   // e.g., "data/sentences.json" or ["data/sentences.json", "data/sentence_expansions/phase5_batch_01.json"]
   masterWordsPath: string;        // e.g., "data/masterWords.json"
   masterSentencesPath: string;    // e.g., "data/masterSentences.json"
